requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
exceptiongroup==1.1.0
execnet==1.9.0
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
packaging==23.0
playwright==1.30.0
pluggy==1.0.0
pyee==9.0.4
pytest==7.2.1
pytest-base-url==2.0.0
pytest-playwright==0.3.0
pytest-xdist==3.2.0
python-dotenv==0.21.1
python-slugify==8.0.1
requests==2.28.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
wincertstore==0.2
